+++ packages/blocks-core/js/libs/wordpress/list-item/index.js
@@ -31,4 +31,18 @@ export const ListItem: BlockType = {
 	edit: (props) => {
 		return <SharedBlockExtension {...props} />;
 	},
+	blockFeatures: {
+		icon: {
+			status: true,
+			inspector: {
+				innerBlocks: {
+					items: {
+						'elements/icon': {
+							availableBlockStates: {},
+						},
+					},
+				},
+			},
+		},
+	},
 };