+++ packages/blocks-core/js/libs/wordpress/navigation-submenu/test/functionality.blocks-navigation-submenu.e2e.cy.js
@@ -38,7 +38,9 @@ describe(
 
 			cy.get('.block-editor-list-view-tree').within(() => {
 				// Open blocks menu
-				cy.get('[aria-label="Add block"]').first().click();
+				cy.get('[aria-label="Add block"]')
+					.first()
+					.click({ force: true });
 			});
 
 			// search for Submenu
@@ -49,7 +51,7 @@ describe(
 			// insert
 			cy.get('button.editor-block-list-item-navigation-submenu')
 				.last()
-				.click();
+				.click({ force: true });
 
 			// enter link value
 			cy.get('input[type="text"]:focus').type('#test{enter}');
@@ -63,6 +65,13 @@ describe(
 			cy.get('.blockera-extension-block-card.master-block-card').should(
 				'exist'
 			);
+
+			// Icon extension is active
+			cy.getByDataTest('settings-tab').click();
+			cy.getByAriaLabel('Choose Icon…').should('be.visible');
+
+			// switch back to style tab
+			cy.getByDataTest('style-tab').click();
 
 			// switch to parent navigation block
 			cy.get('.blockera-extension-block-card.master-block-card').within(