+++ packages/blocks-core/php/libs/wordpress/button/block.php
@@ -14,10 +14,10 @@ return array_merge(
 			$args['selectors'] ?? [],
 			[
 				'blockera/elements/icon' => [
-					'root' => ' .wp-block-button__link .blockera-icon',
+					'root' => 'svg',
 				],
 				'htmlEditable' => [
-					'root' => '.wp-block-button .wp-block-button__link div[role="textbox"][contenteditable="true"], .wp-block-button .wp-block-button__link',
+					'root' => '.wp-block-button .wp-block-button__link',
 				],
 			]
 		),