+++ packages/icons/js/index.js
@@ -6,3 +6,5 @@ export {
 	getIconLibrarySearchData,
 } from './icon-library';
 export { iconSearch } from './icon-search';
+export { default as IconLibrariesList } from './search-libraries.json';
+export { default as NativeIconLibrariesList } from './search-libraries-2.json';